% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_observation_counts.R
\name{add_proj}
\alias{add_proj}
\title{Add project column}
\usage{
add_proj(df)
}
\arguments{
\item{df}{Data.frame with variable called tag_long.}
}
\value{
Data.frame with new colum called proj.
}
\description{
Adds column indicating from which project ad dataset of variable stems if that
information is missing
}
\examples{
df <- add_proj(df) 

}
