% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qog_aggregation_functions.R
\name{agency_bud_to_year_sum}
\alias{agency_bud_to_year_sum}
\title{QoG Agency Fiscal Year to Year}
\usage{
agency_bud_to_year_sum(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to year level for budget variables
}
\description{
Aggregates from qog_agency_year to a year level by summing up the budget variables for each year.
}
\examples{
agency_bud_to_year_sum(df)
}
