% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qog_aggregation_functions.R
\name{agency_inst_to_year}
\alias{agency_inst_to_year}
\title{QoG Agency Instruction to Year}
\usage{
agency_inst_to_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to year level for boolean variables
}
\description{
Aggregates from qog_agency_inst to a year level that can then be translated to country-year.
Groups by year using the year of the agency instruction and calculates the sum of all boolean
variables that have the value 1 in the given year. Hence, the output dataframe is showing how many
agencies were affected by instructions that are e.g. laws.
}
\examples{
agency_inst_to_year(df)

}
