% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdem_aggregation_functions.R
\name{aggregate_vparty_highest_seatshare}
\alias{aggregate_vparty_highest_seatshare}
\title{V-Dem Party-Country-Date to V-Dem Country Year}
\usage{
aggregate_vparty_highest_seatshare(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to country year level
}
\description{
Aggregates from V-Dem Party-Country-Year to Country Year by only keeping the
observations for the largest party per country-year using the variable v2paseatshare
}
\examples{
aggregate_vparty_highest_seatshare(df)

}
