% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{arc}
\alias{arc}
\title{Calculate annualized rate of change (ARC)}
\usage{
arc(years, values, denominator = 100)
}
\arguments{
\item{years}{[\code{numeric()}]\cr
Years corresponding to input values}

\item{values}{[\code{numeric()}]\cr
Input values for some measure evaluated over time}

\item{denominator}{[\code{numeric()}]\cr
Default to 100 for percent (ex: 20\%). Specify denominator = 1 to get
annualized rate of change as a decimal (ex: 0.20).}
}
\value{
[\code{numeric()}]\cr
Annualized rate of change for the intervals between elements of \code{years}.
}
\description{
Calculate annualized rate of change from one vector which
includes numeric year and another vector which includes values of some
measure for those years.
}
\details{
ARC for the period between years i and j is calculated as:
\deqn{arc[i,j] = 100 * ln(value_j / value_i) / (year_j - year_i)}
Index is such that for index i, \code{output[i]} is arc for period \code{years[i - 1]}
to \code{years[i]} and the first entry in the output is \code{NA}.
This function does not sort \code{years} so users should check inputs are
ordered correctly.
}
\examples{
# vectors
arc(years = c(2000, 2005, 2010), values = c(10, 20, 25))

# data.table
library(data.table)
dt <- data.table(
  year = c(2000, 2005, 2010),
  val = c(10, 20, 25)
)
dt[, arc := arc(years = year, values = val)]

}
\seealso{
\itemize{
\item Preston Demography textbook Chapter 1
\item https://www.un.org/esa/sustdev/natlinfo/indicators/methodology_sheets/demographics/population_growth_rate.pdf
}
}
