% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_col_type.R
\name{change_col_type}
\alias{change_col_type}
\title{Change data.table column types}
\usage{
change_col_type(dt, check_cols, is_fun = is.integer, as_fun = as.numeric)
}
\arguments{
\item{dt}{[\code{data.table()}] \cr}

\item{check_cols}{[\code{character()}] \cr
name of columns in \code{dt} to check and modify.}

\item{is_fun}{[\verb{function()}] \cr
function like \code{is.integer()}, \code{is.character()} etc. to check if each of
\code{check_cols} values evaluate to true. Default is \code{is.integer()}.}

\item{as_fun}{[\verb{function()}] \cr
function like \code{as.integer()}, \code{as.character()} etc. to change column values
to. Default is \code{as.numeric()}.}
}
\value{
Modifies \code{dt} in place and returns \code{dt} invisibly.
}
\description{
Checks if \code{check_cols} exist in \code{dt} and are of type \code{is_fun}.
If any of \code{check_cols} are of that  type, then modifies the values using
\code{as_fun}.
}
\examples{
dt <- data.table::data.table(year = 1950L, age_start = 0L)
change_col_type(dt, c("year", "year_start", "age_start"))

}
