% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pdfs.R
\name{combine_pdfs}
\alias{combine_pdfs}
\title{Combine pdfs togethers}
\usage{
combine_pdfs(input_paths, output_path)
}
\arguments{
\item{input_paths}{[\code{character()}] \cr
Existing filepaths to the pdfs that will be combined.}

\item{output_path}{[\code{character(1)}] \cr
Output filepath for the combined pdf to code file. The directory must
exist.}
}
\value{
None. \code{output_path} will be created.
}
\description{
Use PyPDF2 python package to combine multiple pdfs into one pdf.
}
\examples{
\dontrun{
combine_pdfs(
  input_paths = c("diagnostics/USA.pdf", "diagnostics/ZAF.pdf"),
  output_paths = c("diagnostics/combined.pdf")
)
}

}
