% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_prep.R
\name{copy_static_refs}
\alias{copy_static_refs}
\title{copy static reference files}
\usage{
copy_static_refs(
  ROOT_DIR = Sys.getenv("ROOT_DIR"),
  REFS_DIR = file.path("~/proj/demscore/reference_documents"),
  REFS_OUT_DIR = file.path(ROOT_DIR, "autogenerated_refs", "static")
)
}
\arguments{
\item{ROOT_DIR}{The root directory of the project.}

\item{REFS_DIR}{The directory containing the static reference files.}

\item{REFS_OUT_DIR}{The directory to save the references to.}
}
\value{
Returns NULL invisibly.
}
\description{
Copies the static reference files to the autogenerated_refs
directory.
}
