% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_refs.R
\name{create_codebook}
\alias{create_codebook}
\title{Create codebook}
\usage{
create_codebook(args, REF_STATIC_DIR, POSTGRES_TABLES_DIR, LOCAL, prep = FALSE)
}
\arguments{
\item{args}{Command line arguments as a list (usually passed on from docopt)
\itemize{
\item help (flag): Show help message
\item include_unit_cols (flag): Include unit table variables
\item download_id (value): Download id
\item project (value): Project tag
\item dataset (value): Dataset tag
\item thematic (value): Theme tag
\item output_unit_tag (value): Output unit tag
\item variables (value): Variable long tags
\item outfile (value): Output file
\item file_format (value): Output file format (R, CSV, STATA)
\item DJANGO_DB_HOST_DATA (value): Django database host
\item DJANGO_DB_NAME_DATA (value): Django database name
\item DJANGO_DB_USER_DATA (value): Django database user
\item DJANGO_DB_PASSWORD_DATA (value): Django database password
}}

\item{REF_STATIC_DIR}{Path to static codebook files}

\item{POSTGRES_TABLES_DIR}{Path to prepped postgres tables}

\item{LOCAL}{Logical, TRUE if the function should not connect to a database,
FALSE if it should, connection details are passed in \code{args}}
}
\value{
Returns NULL.
}
\description{
Creates the demscore codebook from command line arguments
and directories containing static codebook files and prepped postgres tables.
The appropriate command line arguments and environment variables must be defined.
}
\details{
The codebook can be created in one of four modes, variable selection,
project selection, dataset selection and thematic selection. The mode is
determined by the command line arguments.
}
