% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_methodology.R
\name{create_methodology}
\alias{create_methodology}
\title{Create methodology document}
\usage{
create_methodology(
  REFDIR = file.path("~/proj/demscore/reference_documents/demscoremethodology"),
  temp_dir = tempdir(),
  ROOT_DIR = Sys.getenv("ROOT_DIR"),
  INDIR = file.path(ROOT_DIR, "autogenerated_refs", "refs"),
  outfile =
    file.path("~/proj/demscore/reference_documents/demscoremethodology/methodology.pdf")
)
}
\arguments{
\item{REFDIR}{Path to the directory with the static files (intro.tex etc.).}

\item{temp_dir}{Path to the directory where temporary files are stored.}

\item{ROOT_DIR}{Path to the root directory of the Demscore project.}

\item{INDIR}{Path to the directory with the prepped postgres tables.}

\item{outfile}{Path to the output file.}
}
\description{
Create a pdf document with the methodology of the Demscore project.
}
