% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_prep.R
\name{download_refs}
\alias{download_refs}
\title{Download references from database}
\usage{
download_refs(
  ROOT_DIR = Sys.getenv("ROOT_DIR"),
  OUTDIR = file.path(ROOT_DIR, "autogenerated_refs", "refs"),
  db = pg_connect()
)
}
\arguments{
\item{ROOT_DIR}{The root directory of the project.}

\item{OUTDIR}{The directory to save the references to.}

\item{db}{The database connection.}
}
\value{
Returns NULL invisibly.
}
\description{
Downloads the references from the database and saves them to
a prespecified directory. These files are later used for dataset and
codebook generation.
}
