% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{download_undata_table}
\alias{download_undata_table}
\title{Download data.un.org tables as csvs}
\usage{
download_undata_table(
  table_url,
  output_dir,
  record_limit = 1e+05,
  download_sleep = 30,
  base_url = "http://data.un.org/Data.aspx",
  downloader_url = "http://data.un.org/Handlers/DownloadHandler.ashx"
)
}
\arguments{
\item{table_url}{[\code{character(1)}]\cr
The full url to a specific table. If filters, columns, sort order or other
options are selected, use the 'Link to this page' -> 'Update' button to update
the \code{table_url}.}

\item{output_dir}{[\code{character(1)}]\cr
Directory to save separate downloaded data files.}

\item{record_limit}{[\code{numeric(1)}]\cr
Upper limit of records/rows that you can download at once.}

\item{download_sleep}{[\code{numeric(1)}]\cr
Seconds to sleep/wait between download calls.}

\item{base_url}{[\code{character(1)}]\cr
The base url of the data.un.org site. Default is 'http://data.un.org/Data.aspx'.}

\item{downloader_url}{The base url of the data.un.org site download interface. Default is
'http://data.un.org/Handlers/DownloadHandler.ashx'.}
}
\value{
[\code{character()}] file paths for each separate file downloaded.
}
\description{
Download data.un.org tables as csvs
}
\details{
A list of all possible tables is shown here http://data.un.org/Explorer.aspx.
Currently this function assumes one is able to filter by Year when
downloading data.

The 'data.un.org' website does not allow one to download more than 100,000
records at a time. In order to account for this, we loop through groups of
years that stay below the limit. This means \code{table_url} should not already
include specified filters for years as this function is intended to download
all years of data.
}
\examples{
\dontrun{
table <- download_undata_table(
  table_url = "http://data.un.org/Data.aspx?d=POP&f=tableCode:7;areaCode:0&v=1",
  output_dir = tempdir()
)
}

}
\seealso{
Other get_undata: 
\code{\link{get_undata_nrecords}()},
\code{\link{get_undata_update_dates}()},
\code{\link{undata_url_to_query}()}
}
\concept{get_undata}
