% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{extrapolate}
\alias{extrapolate}
\title{Extrapolate}
\usage{
extrapolate(
  dt,
  id_cols,
  extrapolate_col,
  value_col,
  extrapolate_vals,
  method,
  n_groups_fit
)
}
\arguments{
\item{dt}{[\code{data.table()}]\cr
Data you would like to extrapolate.}

\item{id_cols}{[\code{characher()}]\cr
Columns which uniquely identify rows of \code{dt}. Extrapolation will be
done separately by group on \code{id_cols} excluding \code{interpolate_col}.}

\item{extrapolate_col}{[\code{characher(1)}]\cr
Name of column of \code{dt} which is a numeric variable defining the groups
you would like to extrapolate for.}

\item{value_col}{[\code{characher(1)}]\cr
Name of a column of \code{dt} which is a numeric variable defining the values
you would like to extrapolate.}

\item{extrapolate_vals}{[\code{numeric(1)}]\cr
The values of \code{extrapolate_col} for which you would like to solve for
an extrapolate value of \code{value_col}.}

\item{method}{[\code{characher(1)}]\cr
The method for extrapolation. Must be either 'linear', 'rate_of_change',
or 'uniform'.}

\item{n_groups_fit}{[\code{numeric(1)}]\cr
The number of values of \code{extrapolate_col} to include in determining the
extrapolation. For method 'linear' this is the groups used to fit a linear
model. For method 'rate_of_change' this is the groups used to calculate a
mean rate of change. For method 'uniform' this is the number of groups to
combine into a uniform mean value which will be filled in. The groups will
be selected as the first N groups or the last N groups in the data
depending on the direction of extrapolation. Example: for extrapolation
from years 1990:2000 up to year 2020 with \code{n_groups_fit} = 5, only
years 1995:2000 will be used to fit the extrapolation model.}
}
\value{
[\code{data.table()}] \code{dt} with added rows for extrapolated values.
}
\description{
Extrapolate to dimensions beyond the range of those which
exist in your data.table. Uses linear, rate of change, or uniform
extrapolation.
}
\examples{
dt <- data.table::data.table(
  group = c(rep("a", 5), rep("b", 5)),
  x = rep(c(1:5), 2),
  y = rep(seq(10, 50, 10), 2)
)
dt <- extrapolate(dt, id_cols = c("group", "x"), extrapolate_col = "x",
                  value_col = "y", extrapolate_vals = c(1:10),
                  method = "linear", n_groups_fit = 3)

}
