% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_data.R
\name{generate_data}
\alias{generate_data}
\title{Create Demscore dataset file}
\usage{
generate_data(
  args,
  POSTGRES_TABLES_DIR,
  UNIT_DATA_DIR,
  LOCAL,
  UNIT_TABLE_DIR,
  prep = FALSE
)
}
\arguments{
\item{args}{Command line arguments as a list (usually passed on from docopt)
\itemize{
\item help (flag): Show help message
\item include_unit_cols (flag): Include unit table variables
\item download_id (value): Download id
\item project (value): Project tag
\item dataset (value): Dataset tag
\item thematic (value): Theme tag
\item output_unit_tag (value): Output unit tag
\item variables (value): Variable long tags
\item outfile (value): Output file
\item file_format (value): Output file format (R, CSV, STATA)
\item remove_empty_rows (flag): Remove empty rows.
\item DJANGO_DB_HOST_DATA (value): Django database host
\item DJANGO_DB_NAME_DATA (value): Django database name
\item DJANGO_DB_USER_DATA (value): Django database user
\item DJANGO_DB_PASSWORD_DATA (value): Django database password
\item data_version (value): Data version
}}

\item{POSTGRES_TABLES_DIR}{Path to prepped postgres tables}

\item{UNIT_DATA_DIR}{Path to unit data}

\item{LOCAL}{Logical, TRUE if the function should not connect to a database,
FALSE if it should, connection details are passed in \code{args}}

\item{UNIT_TABLE_DIR}{Path to unit tables}
}
\value{
Returns NULL.
}
\description{
Creates the demscore dataset from command line arguments,
prepped postgres tables, and the unit data and unit tables.
The apropriate command line arguments and environment variables must be defined.
}
\details{
The dataset can be created in one of four modes, variable selection,
project selection, dataset selection and thematic selection. The mode is
determined by the command line arguments.
}
