% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{get_undata_npages}
\alias{get_undata_npages}
\title{Get the total number of pages available in the specified table.}
\usage{
get_undata_npages(bow, query)
}
\arguments{
\item{bow}{host introduction object of class \code{polite}, \code{session} created by \code{bow()} or \code{nod()}}

\item{query}{named list of parameters to be appended to URL in the format \code{list(param1=valA, param2=valB)}}
}
\value{
[\code{numeric(1)}] the total number of pages available in a table as
specified by \code{query}.
}
\description{
Get the total number of pages available in the specified table.
}
\seealso{
Other get_undata: 
\code{\link{get_undata_table}()},
\code{\link{get_undata_update_dates}()},
\code{\link{undata_url_to_query}()}
}
\concept{get_undata}
