% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{get_undata_table_page}
\alias{get_undata_table_page}
\title{Helper function to scrape one page from a table on data.un.org}
\usage{
get_undata_table_page(bow, query, page)
}
\arguments{
\item{bow}{host introduction object of class \code{polite}, \code{session} created by \code{bow()} or \code{nod()}}

\item{query}{named list of parameters to be appended to URL in the format \code{list(param1=valA, param2=valB)}}

\item{page}{[\code{numeric(1)}]\cr
The page to scrape from the specified table.}
}
\description{
Helper function to scrape one page from a table on data.un.org
}
