% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{get_undata_update_dates}
\alias{get_undata_update_dates}
\title{Get the listed update dates for a data.un.org table}
\usage{
get_undata_update_dates(bow, query)
}
\arguments{
\item{bow}{host introduction object of class \code{polite}, \code{session} created by \code{bow()} or \code{nod()}}

\item{query}{named list of parameters to be appended to URL in the format \code{list(param1=valA, param2=valB)}}
}
\value{
a named [\code{Date()}] vector, the names correspond to the description
of the date (likely 'Last' or 'Next').
}
\description{
Get the listed update dates for a data.un.org table
}
\seealso{
Other get_undata: 
\code{\link{download_undata_table}()},
\code{\link{get_undata_nrecords}()},
\code{\link{undata_url_to_query}()}
}
\concept{get_undata}
