% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdata_aggregation_functions.R
\name{hdata_cabinet_date_to_country_date}
\alias{hdata_cabinet_date_to_country_date}
\title{hdata_cabinet_date_to_country_date}
\usage{
hdata_cabinet_date_to_country_date(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to hdata_country_date.
}
\description{
Aggregates from hdata_cabinet_date to hdata_country_date
}
\examples{
hdata_cabinet_date_to_country_date(df)

}
