% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdata_aggregation_functions.R
\name{hdata_country_date_to_country_year_cabinets}
\alias{hdata_country_date_to_country_year_cabinets}
\title{hdata_country_date_to_country_year_cabinets}
\usage{
hdata_country_date_to_country_year_cabinets(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to hdata_country_year.
}
\description{
Aggregates from hdata_country_date to hdata_country_year
}
\examples{
hdata_country_date_to_country_year_cabinets(df)

}
