% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdata_aggregation_functions.R
\name{hdata_dyad_year_to_country_year}
\alias{hdata_dyad_year_to_country_year}
\title{Aggregation hdata_diprep_to_country_year}
\usage{
hdata_dyad_year_to_country_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to hdata_country_year.
}
\description{
Aggregates from dyad_year to country_year by counting the levels of representation for country 1 in a given year.
}
\examples{
hdata_diprep_to_country_year(df)

}
