% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{pct_change}
\alias{pct_change}
\title{Calculate percent change}
\usage{
pct_change(x, y, denominator = 100)
}
\arguments{
\item{x}{[\code{numeric()}]\cr
Baseline value(s)}

\item{y}{[\code{numeric()}]\cr
Comparison value(s)}

\item{denominator}{[\code{numeric()}]\cr
Default to 100 for percent (ex: 20\%). Specify denominator = 1 to get
percent change as a decimal (ex: 0.20).}
}
\value{
[\code{numeric()}]\cr
Percent change from x to y: denominator * (y - x) / x
}
\description{
Calculate percent change from one vector (x) to another (y)
}
\examples{
# vectors
pct_change(x = c(10, 11, 12), y = c(100, 22, 18))

# data.table
library(data.table)
dt <- data.table(
  x = c(10, 11, 12),
  y = c(100, 22, 18)
)
dt[, change := pct_change(x, y)]

# percent change within one data.table column
dt <- data.table(
  year = c(2000, 2001, 2002),
  val = c(10, 11, 12)
)
dt[, change := pct_change(x = shift(val), y = val)]

}
