% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_prep.R
\name{prepare_autogen}
\alias{prepare_autogen}
\title{Prepare references for codebook and download interface}
\usage{
prepare_autogen()
}
\description{
This function prepares the references for the codebook and the download interface.
It downloads and prepares reference files.
}
