% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogen_prep.R
\name{prepare_refs}
\alias{prepare_refs}
\title{Prepare references}
\usage{
prepare_refs(
  ROOT_DIR = Sys.getenv("ROOT_DIR"),
  INDIR = file.path(ROOT_DIR, "autogenerated_refs", "refs"),
  OUTDIR = file.path(ROOT_DIR, "autogenerated_refs", "refs_prepped")
)
}
\arguments{
\item{ROOT_DIR}{The root directory of the project.}

\item{INDIR}{The directory containing the reference files.}

\item{OUTDIR}{The directory to save the references to.}
}
\description{
Prepares the references for use in the autogenerated codebook and dataset
}
