% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{read_undata_files}
\alias{read_undata_files}
\title{Combine together downloaded undata files}
\usage{
read_undata_files(
  fpaths,
  n_footnote_col = 2,
  data_footnoteid_col = "Value Footnotes",
  footnoteid_col = "footnoteSeqID",
  footnote_col = "Footnote"
)
}
\arguments{
\item{fpaths}{[\code{character()}]\cr
Path to separate undata files to combine together. Output of
\code{download_undata_table}.}

\item{n_footnote_col}{[\code{numeric(1)}]\cr
Column number of the 'data' table where the \code{footnote_col} is also stored.
Default is '2'.}

\item{data_footnoteid_col}{[\code{character(1)}]\cr
Name of column in the data table storing the id number. Default is
'Value Footnotes'.}

\item{footnoteid_col}{[\code{character(1)}]\cr
Name of column in the footnote table storing the id number. Default is
'footnoteSeqID'.}

\item{footnote_col}{[\code{character(1)}]\cr
Name of column in the footnote table storing the text. Default is 'Footnote'.}
}
\value{
[\code{data.table(1)}] with all data from input paths combined together.
}
\description{
Combine together downloaded undata files
}
\details{
undata tables contain the standard dataset included in the online tables
plus sometimes a table at the bottom containing footnotes. This function
}
\examples{
\dontrun{
data <- read_undata_files
}
}
\concept{undata}
