% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{read_unit_data}
\alias{read_unit_data}
\title{Read unit data}
\usage{
read_unit_data(unit, ds_tag, variables, datasets, ...)
}
\arguments{
\item{unit}{Unit tag}

\item{ds_tag}{Dataset tag}

\item{variables}{Variables table from postgres}

\item{datasets}{Datasets table from postgres}

\item{...}{Additional arguments to \code{\link{vbase::read_file}}}
}
\value{
Returns a data.frame for variables in the specified dataset from
the specified output unit. The data.frame has as class first the
output unit tag, then the dataset tag, and then "data.frame".
}
\description{
Read data from unit_data/ directory for a
given unit tag and dataset tag. If a variable is not found an warning
is issued. If none of the variables are found, then an error is raised.
}
