% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repdem_aggregation_functions.R
\name{repdem_cabinet_date_to_country_date}
\alias{repdem_cabinet_date_to_country_date}
\title{repdem_cabinet_date_to_country_date}
\usage{
repdem_cabinet_date_to_country_date(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to repdem_country_date.
}
\description{
Aggregates from repdem_cabinet_date to repdem_country_date
}
\examples{
repdem_cabinet_date_to_country_date(df)

}
