% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repdem_aggregation_functions.R
\name{repdem_cabinet_date_to_country_year}
\alias{repdem_cabinet_date_to_country_year}
\title{repdem_cabinet_date_to_country_year}
\usage{
repdem_cabinet_date_to_country_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to repdem_country_year.
}
\description{
Sequentially aggregates from cabinet_date to country_date, then country_date to country_year
}
\examples{
repdem_cabinet_date_to_country_year(df)

}
