% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repdem_aggregation_functions.R
\name{repdem_country_date_to_country_year}
\alias{repdem_country_date_to_country_year}
\title{repdem_country_date_to_country_year}
\usage{
repdem_country_date_to_country_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to repdem_country_year.
}
\description{
Aggregates from repdem_country_date to repdem_country_year
}
\examples{
repdem_country_date_to_country_year(df)

}
