% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_observation_counts.R
\name{reshape_obs}
\alias{reshape_obs}
\title{Reshape observation counts to long format}
\usage{
reshape_obs(df)
}
\arguments{
\item{df}{the observation count data.frame that is to be reshaped.}
}
\value{
A reshaped data.frame in long format.
}
\description{
Reshapes the obervation counts data.frame, that is to be loaded into the globa environment
to a data.frame with one row per variable and output unit. The observations per
variable and output unit are stored in the obs column instead of in one column
for each output unit.
}
\examples{
obs_re <- reshape_obs(obs_counts) 

}
