% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_observation_counts.R
\name{reshape_obs_wide}
\alias{reshape_obs_wide}
\title{Reshape observation counts to wide format}
\usage{
reshape_obs_wide(df)
}
\arguments{
\item{df}{the observation counts data.frame, obs_re, in long format.}
}
\value{
A reshaped data.frame in wide format.
}
\description{
Shapes the reshaped observation counts data frame back to wide format, but with one
column per variable and a unit column storing each output unit.
The fnction calls the reshape_obs function first
}
\examples{
obs_wide <- reshape_obs_wide(obs_counts) 

}
