% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{rinvgamma}
\alias{rinvgamma}
\title{The Inverse Gamma Distribution}
\usage{
rinvgamma(n, shape, scale)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be
the number required.}

\item{shape}{shape parameter of Inverse Gamma distribution, also known as the
alpha parameter. Must be positive.}

\item{scale}{scale parameter of Inverse Gamma distribution, also known as the
beta parameter. Must be positive.}
}
\value{
numeric vector of length \code{n}.
}
\description{
Random generation for the Inverse Gamma distribution with
parameters \code{shape} (alpha) and \code{scale} (beta).
}
