% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{single_download_undata_table}
\alias{single_download_undata_table}
\title{Helper function to download all records for a given data.un.org query}
\usage{
single_download_undata_table(
  query,
  output_path,
  bow,
  record_limit = 1e+05,
  downloader_url = "http://data.un.org/Handlers/DownloadHandler.ashx"
)
}
\arguments{
\item{query}{[\code{list(1)}]\cr
Query with filters, columns, sort order arguments specified as returned by
\code{undata_url_to_query}.}

\item{output_path}{[\code{character(1)}]\cr
Full path and file name to save downloaded csv file to.}

\item{bow}{host introduction object of class \code{polite}, \code{session} created by \code{bow()} or \code{nod()}}

\item{record_limit}{[\code{numeric(1)}]\cr
Upper limit of records/rows that you can download at once.}

\item{downloader_url}{The base url of the data.un.org site download interface. Default is
'http://data.un.org/Handlers/DownloadHandler.ashx'.}
}
\value{
[\code{character(1)}] invisibly returns \code{output_path}where downloaded
file is saved.
}
\description{
Helper function to download all records for a given data.un.org query
}
