% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_hdata_minister_date.R
\name{to_u_hdata_minister_date}
\alias{to_u_hdata_minister_date}
\title{Translation to hdata_minister_date}
\usage{
to_u_hdata_minister_date(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to hdata_minister_date.
}
\description{
Translates compatible dataset to primary unit hdata_minister_date
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_hdata_minister_date(df)

}
