% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_qog_agency_inst.R
\name{to_u_qog_agency_inst}
\alias{to_u_qog_agency_inst}
\title{Translation to u_qog_agency_inst}
\usage{
to_u_qog_agency_inst(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to cucdp_triad_year
}
\description{
Translates compatible dataset to primary unit ucdp_triad_year
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_qog_agency_inst(df)

}
