% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_qog_country.R
\name{to_u_qog_country}
\alias{to_u_qog_country}
\title{Translation to_u_qog_country}
\usage{
to_u_qog_country(df, ...)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame translated to qog_country.
}
\description{
Translates compatible dataset to primary unit u_qog_country.
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_qog_country(df)

}
