% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_qog_region.R
\name{to_u_qog_region}
\alias{to_u_qog_region}
\title{Translation to u_qog_region}
\usage{
to_u_qog_region(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to qog_region.
}
\description{
Translates compatible dataset to primary unit qog_region
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_qog_region(df)

}
