% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_qog_resp_eqi_21.R
\name{to_u_qog_resp_eqi_21}
\alias{to_u_qog_resp_eqi_21}
\title{Translation to u_qog_resp_eqi_21}
\usage{
to_u_qog_resp_eqi_21(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to  u_qog_resp_eqi_21
}
\description{
Translates compatible dataset to primary unit u_qog_resp_eqi_21
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_qog_resp_eqi_21(df)

}
