% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_ucdp_gedid.R
\name{to_u_ucdp_gedid}
\alias{to_u_ucdp_gedid}
\title{Translation to u_ucdp_gedid}
\usage{
to_u_ucdp_gedid(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to cucdp_gedid
}
\description{
Translates compatible dataset to primary unit ucdp_gedid
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_ucdp_gedid(df)

}
