% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_ucdp_pa_dyad_year.R
\name{to_u_ucdp_pa_dyad_year}
\alias{to_u_ucdp_pa_dyad_year}
\title{Translation to u_ucdp_pa_dyad_year}
\usage{
to_u_ucdp_pa_dyad_year(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to cucdp_pa_dyad_year
}
\description{
Translates compatible dataset to primary unit ucdp_pa_dyad_year
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_ucdp_pa_dyad_year(df)

}
