% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_u_vdem_party_date_coder.R
\name{to_u_vdem_party_date_coder}
\alias{to_u_vdem_party_date_coder}
\title{Translation to vdem_party_date_coder}
\usage{
to_u_vdem_party_date_coder(df, ...)
}
\arguments{
\item{df}{a data.frame.}
}
\value{
data.frame translated to vdem_party_date_coder.
}
\description{
Translates compatible dataset to primary unit vdem_party_date_coder
}
\section{Note}{
 Unit tables are not passed as function arguments, but are automatically loaded from the \code{ROOT} directory
}

\examples{
to_u_vdem_party_date_coder(df)

}
