% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucdp_aggregation_functions.R
\name{u_ucdp_gedid_to_country_year}
\alias{u_ucdp_gedid_to_country_year}
\title{ucdp_gedid to country_year}
\usage{
u_ucdp_gedid_to_country_year(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
data.frame aggregated to ucdp_country_year for fatality variables
for each type of violence.
}
\description{
Aggregates from ucdp_gedid to country_year by using the fatality variables
in the ucdp_ged dataset. The fatalities are aggregated by country, year and type of violence (onesided, nonstate, statebased), which are mutually exclusive.
Variables not counting/estimating fatalities are excluded from the aggregation. Those variabes are:
are "relid","code_status","type_of_violence","conflict_dset_id",
"conflict_new_id","conflict_name","dyad_dset_id","dyad_new_id","dyad_name",
"side_a_dset_id","side_a_new_id", "side_a","side_b_dset_id", "side_b_new_id","side_b",
"number_of_sources","source_article", "source_office", "source_date", "source_headline"
,"source_original", "where_prec","where_coordinates", "where_description", "adm_1",
"adm_2","latitude", "longitude", "geom_wkt", "priogrid_gid", "country", "country_id",
"region", "event_clarity","date_prec", "date_start","date_end", "gwnoa", "gwnob"
}
\examples{
u_ucdp_gedid_to_country_year(df)

}
