% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undata.R
\name{undata_url_to_query}
\alias{undata_url_to_query}
\title{Convert the undata url to the query argument needed to reference
specific tables}
\usage{
undata_url_to_query(table_url, base_url = "http://data.un.org/Data.aspx")
}
\arguments{
\item{table_url}{[\code{character(1)}]\cr
The full url to a specific table. If filters, columns, sort order or other
options are selected, use the 'Link to this page' -> 'Update' button to update
the \code{table_url}.}

\item{base_url}{[\code{character(1)}]\cr
The base url of the data.un.org site. Default is 'http://data.un.org/Data.aspx'.}
}
\value{
[\code{list()}] of query parameters to pass to \code{get_undata_table} or
other related undata functions.
}
\description{
Convert the undata url to the query argument needed to reference
specific tables
}
\examples{
table_url <- "http://data.un.org/Data.aspx?d=PopDiv&f=variableID\%3a12"
query <- undata_url_to_query(table_url)

}
\seealso{
Other get_undata: 
\code{\link{get_undata_npages}()},
\code{\link{get_undata_table}()},
\code{\link{get_undata_update_dates}()}
}
\concept{get_undata}
