% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invgamma.R
\name{rinvgamma}
\alias{rinvgamma}
\title{The Inverse Gamma Distribution}
\usage{
rinvgamma(n, shape, scale)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, the length
    is taken to be the number required.}

\item{shape}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}

\item{scale}{shape and scale parameters.  Must be positive,
    \code{scale} strictly.}
}
\value{
numeric vector of length \code{n}.
}
\description{
Random generation for the Inverse Gamma distribution with
parameters \code{shape} (alpha) and \code{scale} (beta).
}
\examples{
rinvgamma(10, 0.5, 0.01)

}
